Refactor code to pass CI
