+++ src/test/java/listpackage/CourseListTest.java
@@ -1,4 +1,4 @@
-package TutorLink.listpackage;
+package listpackage;
 
 public class CourseListTest {
 }
